Use f32 as main type

+++ src/mp3/tests.zig
@@ -54,7 +54,7 @@ test "MP3 info" {
 
     try testing.expectEqual(@as(u32, 44100), info.sample_rate);
     try testing.expectEqual(@as(u8, 2), info.channels);
-    try testing.expectEqual(SampleType.i16, info.sample_type);
+    try testing.expectEqual(SampleType.f32, info.sample_type);
     try testing.expect(info.total_frames > 0);
 }
 
@@ -69,10 +69,10 @@ test "MP3 open + streaming" {
 
     try testing.expectEqual(@as(u32, 44100), decoder.info.sample_rate);
     try testing.expectEqual(@as(u8, 2), decoder.info.channels);
-    try testing.expectEqual(SampleType.i16, decoder.info.sample_type);
+    try testing.expectEqual(SampleType.f32, decoder.info.sample_type);
     try testing.expect(decoder.info.total_frames > 0);
 
-    var buf: [2048]i16 = undefined;
+    var buf: [2048]f32 = undefined;
     const read = try decoder.read(&buf);
     try testing.expect(read > 0);
 }
@@ -83,10 +83,10 @@ test "MP3 decodeMemory" {
 
     try testing.expectEqual(@as(u32, 44100), audio.params.sample_rate);
     try testing.expectEqual(@as(u8, 2), audio.params.channels);
-    try testing.expectEqual(SampleType.i16, audio.params.sample_type);
+    try testing.expectEqual(SampleType.f32, audio.params.sample_type);
     try testing.expect(audio.data.len > 0);
 
-    const frame_size = audio.params.channels * @sizeOf(i16);
+    const frame_size = audio.params.channels * @sizeOf(f32);
     try testing.expect(audio.data.len % frame_size == 0);
 }
